completed Prefix usage for property filters and added some ExceptionHandling
