+++ lib/jekyll-rdf.rb
@@ -38,7 +38,6 @@ require 'jekyll/drops/rdf_resource'
 require 'jekyll/drops/rdf_resource_class'
 require 'jekyll/filters/rdf_sparql_query'
 require 'jekyll/filters/rdf_property'
-require 'jekyll/filters/rdf_property_list'
 require 'jekyll/rdf_main_generator'
 require 'jekyll/rdf_page_data'
 require 'jekyll/rdf_template_mapper'
