+++ lib/jekyll/filters/rdf_property.rb
@@ -40,6 +40,7 @@ module Jekyll
     def rdf_property(input, predicate, lang = nil)
       return input unless input.is_a?(Jekyll::Drops::RdfResource)
       begin
+        predicate = rdf_resolve_prefix(input, predicate)
         results = input.page.data['rdf'].statements_as_subject.select{ |s| s.predicate.term.to_s == predicate }
         lang ||= input.site.config['jekyll_rdf']['language']
         if results.count > 1 && lang != nil
@@ -57,6 +58,45 @@ module Jekyll
       end
     end
 
+    ##
+    # Computes all objects for which statements exist containing the given subject and predicate and returns an Array of them
+    #
+    # * +input+ - is the subject of the statements to be matched
+    # * +predicate+ - is the predicate of the statements to be matched
+    # * +lang+ - (optional) preferred language of the returned objects. If 'cfg' is specified the preferred language is provides by the site configuration _config.yml
+    #
+    def rdf_property_list(input, predicate, lang = nil)
+      return input unless input.is_a?(Jekyll::Drops::RdfResource)
+      begin
+        predicate = rdf_resolve_prefix(input, predicate)
+        result = input.page.data['rdf'].statements_as_subject.select{ |s| s.predicate.term.to_s == predicate } # select all matching statements with given predicate
+        if lang != nil
+          if lang == 'cfg'
+            lang = input.site.config['jekyll_rdf']['language']
+          end
+          result = result.select{ |s|
+            if(s.object.term.is_a?(RDF::Literal))
+              s.object.term.language == lang.to_sym
+            else
+              false
+            end
+          } # select all statements with matching language
+        end
+        return unless result
+        result.map{|p| p.object.name}
+      end
+    end
+    
+    private
+    def rdf_resolve_prefix(input, predicate)
+      if(!input.page.data["rdf_prefixes"].nil?)
+        arr=predicate.split(":",2)  #bad regex, would also devide 'http://example....' into 'http' and '//example....',even though it is already a complete URI; if 'PREFIX http: <http://...> is defined, 'http' in 'http://example....' could be mistaken for a prefix 
+        if(!input.page.data["rdf_prefix_map"][arr[0]].nil?) 
+          return arr[1].prepend(input.page.data["rdf_prefix_map"][arr[0]])
+        end 
+      end
+      return predicate
+    end
   end
 end
