--- lib/jekyll/filters/rdf_property_list.rb
@@ -1,64 +1,0 @@
-##
-# MIT License
-#
-# Copyright (c) 2016 Elias Saalmann, Christian Frommert, Simon Jakobi,
-# Arne Jonas Präger, Maxi Bornmann, Georg Hackel, Eric Füg
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-#
-
-module Jekyll
-
-  ##
-  # Internal module to hold the method #rdf_property
-  #
-  module RdfPropertyList
-
-    ##
-    # Computes all objects for which statements exist containing the given subject and predicate and returns an Array of them
-    #
-    # * +input+ - is the subject of the statements to be matched
-    # * +predicate+ - is the predicate of the statements to be matched
-    # * +lang+ - (optional) preferred language of the returned objects. If 'cfg' is specified the preferred language is provides by the site configuration _config.yml
-    #
-    def rdf_property_list(input, predicate, lang = nil)
-      return input unless input.is_a?(Jekyll::Drops::RdfResource)
-      begin
-        result = input.page.data['rdf'].statements_as_subject.select{ |s| s.predicate.term.to_s == predicate } # select all matching statements with given predicate
-        if lang != nil
-          if lang == 'cfg'
-            lang = input.site.config['jekyll_rdf']['language']
-          end
-          result = result.select{ |s|
-            if(s.object.term.is_a?(RDF::Literal))
-              s.object.term.language == lang.to_sym
-            else
-              false
-            end
-          } # select all statements with matching language
-        end
-        return unless result
-        result.map{|p| p.object.name}
-      end
-    end
-
-  end
-end
-
-Liquid::Template.register_filter(Jekyll::RdfPropertyList)
