+++ lib/jekyll/filters/rdf_sparql_query.rb
@@ -60,7 +60,14 @@ module Jekyll
           hsh.collect{|k,v| [k.to_s, v]}.to_h
         end
         return result
+      rescue SPARQL::Client::ClientError => ce
+        Jekyll.logger.error("malformed query found: \n #{query} \n Error Message: #{ce.message}")
+      rescue SPARQL::MalformedQuery => mq
+        Jekyll.logger.error("malformed query found: \n #{query} \n Error Message: #{mq.message}")
+      rescue Exception => e
+        Jekyll.logger.error("unknown Exception of class: #{e.class} in sparql_query \n Query: #{query} \nMessage: #{e.message}") 
       end
+      return []
     end
 
   end
