+++ lib/jekyll/rdf_page_data.rb
@@ -50,10 +50,14 @@ module Jekyll
       self.data['rdf'] = resource
       if !self.data["rdf_prefix_path"].nil?
         begin
-          prefixFile=File.new(File.join(base, 'rdf-data', self.data["rdf_prefix_path"].strip))
-          self.data["rdf_prefixes"]=prefixFile.readlines.join(" ")
+          prefixFile=File.new(File.join(base, 'rdf-data', self.data["rdf_prefix_path"].strip)).readlines
+          self.data["rdf_prefixes"] = prefixFile.join(" ")
+          self.data["rdf_prefix_map"] = Hash[ *(prefixFile.collect { |v|
+            arr = v.split(":",2)
+            [arr[0][7..-1].strip, arr[1].strip[1..-2]]
+          }.flatten)] 
         rescue Errno::ENOENT => ex
-          Jekyll.logger.warn("context: #{resource.name}  template: #{template}  file not found: #{File.join(base, 'rdf-data', self.data["rdf_prefix_path"])}") 
+          Jekyll.logger.error("context: #{resource.name}  template: #{template}  file not found: #{File.join(base, 'rdf-data', self.data["rdf_prefix_path"])}") 
         end
       end 
       resource.page = self
